requirements.txt
web3==5.25.0
python-dotenv==0.19.2
requests==2.26.0
